claim rewards + more

+++ server.js
@@ -10,10 +10,12 @@ const jwt = require("jsonwebtoken");
 const port = process.env.PORT || 3000;
 const cors = require("cors");
 const helmet = require("helmet");
-// const cookieParser = require("cookie-parser");
+const cookieParser = require("cookie-parser");
 
 // const { fromServer } = require("./endpoint-handler.js");
-const { database } = require("./database.js");
+// const { database } = require("./database.js");
+const { database } = require("./endpoint-resolver.js");
+// const { localStore } = require("./localStore.js");
 const auth = require("./middleware/auth.js");
 const { constants } = require("./constants.js");
 
@@ -30,16 +32,26 @@ const { constants } = require("./constants.js");
 // };
 
 app.use(helmet());
-// app.use(cookieParser());
+app.use(cookieParser());
 app.use(express.json({ limit: "50mb" }));
 
 app.use(express.static(path.join(__dirname, "dist")));
 
 app.use(
   cors({
-    origin: "http://localhost:5173",
+    origin: "http://127.0.0.1:5174",
   })
 );
+
+app.post("/api/auth", auth, async (req, res) => {
+  const token =
+    req.body.token ||
+    req.query.token ||
+    req.params.token ||
+    req.headers["x-access-token"] ||
+    req.cookies.token;
+  res.status(200).send({ token });
+});
 
 app.post("/api/register", async (req, res) => {
   try {
@@ -62,17 +74,21 @@ app.post("/api/register", async (req, res) => {
       email: email.toLowerCase(),
       password: encryptedPassword,
       balance: 100,
+      claimExpiresTimestamp: 0,
     });
 
     const token = jwt.sign({ _userid, username }, process.env.TOKEN_KEY, {
-      expiresIn: "2h",
+      expiresIn: "1h",
     });
 
     await database.addSession(token);
 
-    return res.status(201).json({ token });
-  } catch (error) {
-    console.log(error);
+    return res
+      .cookie("token", token, { httpOnly: true })
+      .status(201)
+      .json({ token });
+  } catch (errors) {
+    console.log(errors);
   }
 });
 
@@ -83,7 +99,9 @@ app.post("/api/login", async (req, res) => {
     if (!(username && password)) {
       res.status(400).send("All input are required");
     }
-
+    if (!username.includes("@")) {
+      res.status(400).send("cannot login with username");
+    }
     const user = await database.getUser(
       username.includes("@") ? "email" : "username",
       username
@@ -97,20 +115,21 @@ app.post("/api/login", async (req, res) => {
       { _userid: user._id, username: user.username },
       process.env.TOKEN_KEY,
       {
-        expiresIn: "2h",
+        expiresIn: "1h",
       }
     );
     await database.addSession(token);
 
-    console.log(token);
-
-    return res.status(200).json({ token });
-  } catch (error) {
-    console.log(error);
+    return res
+      .cookie("token", token, { httpOnly: true })
+      .status(200)
+      .send({ token });
+  } catch (errors) {
+    console.log(errors);
   }
 });
 
-app.post("/api/logout", auth, async (req, res) => {
+app.post("/api/logout", async (req, res) => {
   try {
     const token =
       req.body.token ||
@@ -118,10 +137,13 @@ app.post("/api/logout", auth, async (req, res) => {
       req.params.token ||
       req.headers["x-access-token"];
 
+    if (!(await database.getSession(token))) {
+      return res.status(400).send("failed to find session");
+    }
     const removed = await database.removeSession(token);
 
     if (removed.deletedCount > 0) {
-      return res.status(200).send("logout");
+      return res.status(200).clearCookie("token").send("logout");
     }
     return res.status(400).send("failed to logout");
   } catch (errors) {
@@ -143,8 +165,8 @@ app.get("/api/get-user", auth, async (req, res) => {
       username: user.username,
       balance: user.balance,
     });
-  } catch (error) {
-    console.log(error);
+  } catch (errors) {
+    console.log(errors);
   }
 });
 
@@ -157,8 +179,8 @@ app.get("/api/get-username", auth, async (req, res) => {
     res.status(200).send({
       username: user.username,
     });
-  } catch (error) {
-    console.log(error);
+  } catch (errors) {
+    console.log(errors);
   }
 });
 
@@ -171,8 +193,8 @@ app.get("/api/get-balance", auth, async (req, res) => {
     res.status(200).send({
       balance: user.balance,
     });
-  } catch (error) {
-    console.log(error);
+  } catch (errors) {
+    console.log(errors);
   }
 });
 
@@ -199,35 +221,23 @@ app.get("/api/game/:id", auth, async (req, res) => {
       return res.status(403).send("Unathorised access to game");
     }
 
-    return res.status(200).send(game);
-  } catch (error) {
-    console.log(error);
-  }
-});
-
-app.get("/api/latest-game", auth, async (req, res) => {
-  try {
-    const { _userid } = req.user;
-
-    const game = await database.getUserGames(_userid);
-
-    if (game.length === 0) {
-      return res.status(404).send("No games found");
+    if (game.state !== "done") {
+      delete game.bomb.position;
     }
 
-    return res.status(200).send(game[game.length - 1]);
-  } catch (error) {
-    console.log(error);
+    return res.status(200).send(game);
+  } catch (errors) {
+    console.log(errors);
   }
 });
 
-app.post("/api/balance", auth, async (req, res) => {
+app.get("/api/game/:id/bomb-locations", auth, async (req, res) => {
   try {
     const { _userid } = req.user;
-    const { balance, gameid } = req.body;
+    const gameid = req.params.id;
 
-    if (!(balance && gameid) && balance != 0) {
-      return res.status(400).send("Missing balance or game id");
+    if (!gameid) {
+      res.status(400).send("ID must be provided");
     }
 
     if (gameid.length != 24) {
@@ -243,26 +253,180 @@ app.post("/api/balance", auth, async (req, res) => {
     if (game.belongsTo.toString() !== _userid) {
       return res.status(403).send("Unathorised access to game");
     }
-    if (game.state !== constants.DONE) {
-      return res.status(400).send("Game isn't finished");
+
+    return res.status(200).send(game.bomb.position);
+  } catch (errors) {
+    console.log(errors);
+  }
+});
+
+/*
+when getting game check condiotion of game, if the game is over send everything
+if the game is still ongoing send only necessary information.
+check game state on click and stuff
+
+create game generates NextClick
+
+get game click too
+*/
+app.get("/api/get-claim", auth, async (req, res) => {
+  try {
+    const { _userid } = req.user;
+
+    const user = await database.getUser("_id", _userid);
+
+    if (!user) {
+      return res.status(404).send("User not found");
+    }
+    if (user.claimExpiresTimestamp < new Date()) {
+      return res.status(200).send({
+        claimable: true,
+        claimAmount: 50,
+        timestamp: user.claimExpiresTimestamp,
+      });
+    }
+    return res
+      .status(200)
+      .send({ claimable: false, timestamp: user.claimExpiresTimestamp });
+  } catch (errors) {
+    console.log(errors);
+  }
+});
+
+app.post("/api/claim-reward", auth, async (req, res) => {
+  try {
+    const { _userid } = req.user;
+
+    const user = await database.getUser("_id", _userid);
+    if (!user) {
+      return res.status(404).send("User not found");
     }
 
-    if (game.hasClaimed) {
+    if (user.claimExpiresTimestamp < new Date()) {
+      const date = await database.updateUserClaim(_userid);
+      if (date.modifiedCount === 0) {
+        res.status(400).send("Failed to update user claim");
+      }
+      const updatedBalance = await database.updateBalance(
+        _userid,
+        user.balance + 50
+      );
+      if (updatedBalance.modifiedCount === 0) {
+        res.status(400).send("Failed to update user balance");
+      }
+      const _user = await database.getUser("_id", _userid);
+      return res.status(200).send({
+        claimable: false,
+        timestamp: _user.claimExpiresTimestamp,
+        balance: _user.balance,
+      });
+    }
+    return res
+      .status(200)
+      .send({ claimable: false, timestamp: user.claimExpiresTimestamp });
+  } catch (errors) {
+    console.log(errors);
+  }
+});
+
+app.get("/api/latest-game", auth, async (req, res) => {
+  try {
+    const { _userid } = req.user;
+
+    const game = await database.getUserGames(_userid);
+
+    if (game.length === 0) {
+      return res.status(404).send("No games found");
+    }
+    const latestGame = game[game.length - 1];
+    if (latestGame.state !== "done") {
+      delete latestGame.bomb.position;
+    }
+    return res.status(200).send(latestGame);
+  } catch (errors) {
+    console.log(errors);
+  }
+});
+
+app.get("/api/game-history", auth, async (req, res) => {
+  try {
+    const { _userid } = req.user;
+
+    const game = await database.getUserGames(_userid);
+
+    if (game.length === 0) {
+      return res.status(404).send("No games found");
+    }
+
+    console.log(game);
+
+    const gameHistory = game.filter((value) => value.state === constants.DONE);
+
+    // if (latestGame.state !== "done") {
+    //   delete latestGame.bomb.position;
+    // }
+    return res.status(200).send(gameHistory);
+  } catch (errors) {
+    console.log(errors);
+  }
+});
+
+app.post("/api/balance", auth, async (req, res) => {
+  try {
+    const { _userid } = req.user;
+    const { balance, gameid } = req.body;
+    console.log("exist");
+    if (!(balance && gameid) && balance !== 0) {
+      return res.status(400).send("Missing balance or game id");
+    }
+    console.log("id");
+
+    if (gameid.length != 24) {
+      return res.status(400).send("A valid ID is required");
+    }
+
+    const game = await database.getGame(gameid);
+    console.log("game");
+
+    if (!game) {
+      return res.status(404).send("Game does not exist");
+    }
+    console.log("userid");
+
+    if (game.belongsTo.toString() !== _userid) {
+      return res.status(403).send("Unathorised access to game");
+    }
+    //
+    console.log("done");
+
+    if (balance > 0 && game.state !== constants.DONE) {
+      return res.status(400).send("Game isn't finished");
+    }
+    console.log(1);
+    console.log("has");
+
+    if (balance > 0 && game.hasClaimed) {
       return res.status(400).send("Game has already been claimed");
     }
+
+    console.log("set");
 
     const user = await database.getUser("_id", _userid);
     const newBalance = balance + user.balance;
     const updatedBalance = await database.updateBalance(_userid, newBalance);
 
     if (updatedBalance.modifiedCount > 0) {
-      await database.updateClaimedGame(gameid);
+      console.log("pass");
+      if (balance > 0) {
+        await database.updateClaimedGame(gameid);
+      }
       return res.status(200).send({ balance: newBalance });
     }
+    console.log("failed");
 
     return res.status(400).send("Failed to update");
-  } catch (error) {
-    console.log(error);
+  } catch (errors) {
+    console.log(errors);
   }
 });
 
@@ -278,10 +442,10 @@ app.get("/api/game/:id/click", auth, async (req, res) => {
   try {
     const { _userid } = req.user;
     const gameid = req.params.id;
-    const { clickPosition } = req.body;
+    const clickPosition = parseInt(req.query.clickPosition);
 
     if (!(gameid && clickPosition) && clickPosition != 0) {
-      res.status(400).send("Missing game id or click position");
+      return res.status(400).send("Missing game id or click position");
     }
 
     if (gameid.length != 24) {
@@ -306,13 +470,43 @@ app.get("/api/game/:id/click", auth, async (req, res) => {
       return res.status(400).send("Can't click the same square");
     }
 
-    const click = await database.updateClicks(gameid, clickPosition);
+    let ongoingData = {};
+    let earned = 0;
+    if (game.state !== constants.DONE) {
+      const chanceOFWinning = getPercentageOfWining(
+        game.size,
+        game.clicks.length + 1,
+        game.bomb.count
+      );
+      const pool = (1 / chanceOFWinning) * game.stake;
+      earned = pool - game.pool;
+      console.log(earned);
+
+      const newPool = await database.updateGamePool(gameid, earned);
+      if (newPool !== pool) {
+        return res.status(400).send("miss match in game pools");
+      }
+      const potentialPool = getPercentageOfWining(
+        game.size,
+        game.clicks.length + 2,
+        game.bomb.count
+      );
+      const nextClick = (1 / potentialPool) * game.stake - newPool;
+
+      ongoingData = { nextClick, pool: newPool };
+    }
+
+    const click = await database.addClick(gameid, {
+      position: clickPosition,
+      earned: earned,
+    });
 
     if (click.modifiedCount === 0) {
       return res.status(400).send("failed to update click");
     }
 
     let _game = await database.getGame(gameid);
+    const clickCount = await database.getClicks(gameid);
 
     if (game.bomb.position.includes(clickPosition)) {
       _game = await database.updateGameState(
@@ -320,10 +514,7 @@ app.get("/api/game/:id/click", auth, async (req, res) => {
         constants.DONE,
         constants.LOST
       );
-    } else if (
-      (await database.getClicks(gameid)).length ===
-      game.size - game.bomb.count
-    ) {
+    } else if (clickCount.length === game.size - game.bomb.count) {
       _game = await database.updateGameState(
         gameid,
         constants.DONE,
@@ -331,38 +522,70 @@ app.get("/api/game/:id/click", auth, async (req, res) => {
       );
     }
 
-    let ongoingData = {};
-    if (_game.state !== constants.DONE) {
-      const chanceOFWinning = getPercentageOfWining(
-        _game.size,
-        _game.clicks.length,
-        _game.bomb.count
+    if (_game.state === "done") {
+      ongoingData = { pool: _game.pool - earned };
+      await database.updateGamePool(gameid, -earned);
+    } else {
+      const updateNextClick = await database.updateNextClick(
+        gameid,
+        ongoingData.nextClick
       );
-      const pool = (1 / chanceOFWinning) * _game.stake;
-      const earned = pool - _game.pool;
-
-      const newPool = await database.updateGamePool(gameid, earned);
-      if (newPool !== pool) {
-        return res.status(400).send("miss match in game pools");
+      if (!updateNextClick) {
+        return res.status(400).send("failed to add next click");
       }
-      const potentialPool = getPercentageOfWining(
-        _game.size,
-        _game.clicks.length + 1,
-        _game.bomb.count
-      );
-      const nextClick = (1 / potentialPool) * _game.stake - newPool;
-
-      ongoingData = { earned, nextClick, pool: newPool };
     }
 
-    res.status(200).send({
+    return res.status(200).send({
       state: _game.state,
       result: _game.result,
       pool: _game.pool,
+      clicks: _game.clicks,
       ...ongoingData,
     });
-  } catch (error) {
-    console.log(error);
+  } catch (errors) {
+    console.log(errors);
+  }
+});
+
+app.post("/api/claim-game", auth, async (req, res) => {
+  try {
+    const { _userid } = req.user;
+    const { gameid } = req.body;
+
+    if (!gameid) {
+      return res.status(400).send("Missing game id");
+    }
+
+    if (gameid.length != 24) {
+      return res.status(400).send("A valid ID is required");
+    }
+
+    const game = await database.getGame(gameid);
+
+    if (!game) {
+      return res.status(404).send("Game does not exist");
+    }
+
+    if (game.belongsTo.toString() !== _userid) {
+      return res.status(403).send("Unathorised access to game");
+    }
+
+    if (game.state === "done") {
+      return res.status(400).send("Game already claimed");
+    }
+    console.log(game);
+    if (game.clicks.length === 0) {
+      return res.status(400).send("Must click at least once");
+    }
+
+    const _game = await database.updateGameState(
+      gameid,
+      constants.DONE,
+      constants.CLAIMED
+    );
+    return res.status(200).send(_game);
+  } catch (errors) {
+    console.log(errors);
   }
 });
 
@@ -382,7 +605,7 @@ function generateBombLocation(bombCount) {
         newValue = randomIntFromInterval(1, 25);
       }
       usedValues.push(newValue);
-      return newValue - 1;
+      return newValue;
     })
     .sort((a, b) => 0.5 - Math.random());
 }
@@ -391,16 +614,24 @@ app.post("/api/create-game", auth, async (req, res) => {
   const { _userid } = req.user;
   const { bombCount, stake } = req.body;
 
+  console.log(bombCount, stake);
+  if (!(bombCount && stake)) {
+    return res.status(400).send("Missing bombcount or stake");
+  }
+  console.log("test");
+
   const usergames = await database.getUserGames(_userid);
-  const ongoingGame = usergames.filter((game) => {
-    game.state === constants.ONGOING;
-  });
 
-  console.log(ongoingGame);
+  const ongoingGame = usergames.filter(
+    (game) => game.state === constants.ONGOING
+  );
 
-  if (ongoingGame) {
+  if (ongoingGame.length > 0) {
     return res.status(400).send("User already in a game");
   }
+  const potentialPool = getPercentageOfWining(25, 1, bombCount);
+
+  const nextClick = (1 / potentialPool) * stake - stake;
 
   const game = {
     belongsTo: _userid,
@@ -413,11 +644,12 @@ app.post("/api/create-game", auth, async (req, res) => {
     stake,
     clicks: [],
     pool: stake,
+    nextClick,
   };
 
   const gameid = await database.createGame(game);
 
-  res.status(200).send({ gameid });
+  res.status(200).send(gameid);
 });
 
 http.listen(port, function () {
